Per project Per tasklist permissions

+++ database/migrations/2016_05_02_130907_alter_is_finished_column_task_checklist.php
@@ -3,19 +3,20 @@
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
 
-class AlterIsFinishedColumnTaskChecklist extends Migration
-{
+class AlterIsFinishedColumnTaskChecklist extends Migration {
+
     /**
      * Run the migrations.
      *
      * @return void
      */
-    public function up()
-    {
-         Schema::table('task_check_list', function (Blueprint $table) {
+    public function up() {
+        if (!Schema::hasColumn('status')) {
+            Schema::table('task_check_list', function (Blueprint $table) {
                 $table->dropColumn('is_finished');
                 $table->string('status')->default('Default');
-        });
+            });
+        }
     }
 
     /**
@@ -23,11 +24,11 @@ class AlterIsFinishedColumnTaskChecklist extends Migration
      *
      * @return void
      */
-    public function down()
-    {
+    public function down() {
         Schema::table('task_check_list', function (Blueprint $table) {
-                $table->dropColumn('status');
-                $table->tinyInteger('is_finished');
+            $table->dropColumn('status');
+            $table->tinyInteger('is_finished');
         });
     }
+
 }
