+++ database/migrations/2016_05_11_160936_update_events_table.php
@@ -3,19 +3,20 @@
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
 
-class UpdateEventsTable extends Migration
-{
+class UpdateEventsTable extends Migration {
+
     /**
      * Run the migrations.
      *
      * @return void
      */
-    public function up()
-    {
-        Schema::table('events',function(Blueprint $table){
-            $table->dropColumn('username');
-            $table->integer('user_id')->after('event_id');
-        });
+    public function up() {
+        if (!Schema::hasColumn('user_id')) {
+            Schema::table('events', function(Blueprint $table) {
+                $table->dropColumn('username');
+                $table->integer('user_id')->after('event_id');
+            });
+        }
     }
 
     /**
@@ -23,11 +24,11 @@ class UpdateEventsTable extends Migration
      *
      * @return void
      */
-    public function down()
-    {
-        Schema::table('events',function(Blueprint $table){
+    public function down() {
+        Schema::table('events', function(Blueprint $table) {
             $table->dropColumn('user_id');
             $table->string('username')->after('event_id');
         });
     }
+
 }
